feat: add validation rules
