+++ src/routes/graphql/index.ts
@@ -1,7 +1,7 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
-import { createGqlResponseSchema, gqlResponseSchema } from './schemas.js';
-import { buildSchema, graphql } from 'graphql';
-import { UUIDType } from './types/uuid.js';
+import { createGqlResponseSchema, gqlResponseSchema, gqlSchema } from './schemas.js';
+import { graphql, parse, validate } from 'graphql';
+import depthLimit from 'graphql-depth-limit';
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   const { prisma } = fastify;
@@ -17,8 +17,11 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
     },
     async handler(req) {
       //console.log('зашли в handler graphql !!!!!!!')
+      const errors = validate(gqlSchema, parse(req.body.query), [depthLimit(5)]);
+      if (errors.length) return { errors }; 
+      
       const result = graphql({
-        schema,
+        schema: gqlSchema,
         source: req.body.query,
         variableValues: req.body.variables,
         rootValue: resolvers,
@@ -202,130 +205,5 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   //}
 
 };
-
-const schema = buildSchema(`
-  type Query {
-    memberTypes: [MemberType]
-    posts: [Post]
-    users: [UserProfile]
-    profiles: [ProfileMemberType]
-
-    memberType(id: MemberTypeId!): MemberType
-    post(id: UUID!): Post
-    user(id: UUID!): UserProfile
-    profile(id: UUID!): ProfileMemberType
-  }
-
-  type MemberType {
-    id: MemberTypeId!
-    discount: Float
-    postsLimitPerMonth: Int
-  }
-
-  type Post {
-    id: UUID!
-    title: String
-    content: String
-  }
-
-  type User {
-    id: UUID!
-    name: String
-    balance: Float
-  }
-
-  type Profile {
-    id: UUID!
-    isMale: Boolean
-    yearOfBirth: Int
-  }
-
-  type UserProfile {
-    id: UUID
-    name: String
-    balance: Float
-    profile: ProfileMemberType
-    posts: [Post]
-    userSubscribedTo: [UserSubscribedTo]
-    subscribedToUser: [SubscribedToUser]
-  }
-
-  type ProfileMemberType {
-    id: UUID!
-    isMale: Boolean
-    yearOfBirth: Int
-    memberType: MemberType
-  }
-
-  type UserSubscribedTo {
-    id: UUID!
-    name: String
-    subscribedToUser: [SubscribedToUser]
-  }
-
-  type SubscribedToUser {
-    id: UUID!
-    name: String
-    userSubscribedTo: [UserSubscribedTo]
-  }
-
-  enum MemberTypeId {
-    basic
-    business
-  }
-
-  scalar UUID
-
-  type Mutation  {
-    createUser(dto: CreateUserInput!): User
-    createPost(dto: CreatePostInput!): Post
-    createProfile(dto: CreateProfileInput!): Profile
-
-    deletePost(id: UUID): Boolean
-    deleteProfile(id: UUID): Boolean
-    deleteUser(id: UUID): Boolean
-
-    changeUser(id: UUID, dto: ChangeUserInput!): User
-    changePost(id: UUID, dto: ChangePostInput!): Post
-    changeProfile(id: UUID, dto: ChangeProfileInput!): Profile
-
-    subscribeTo(userId: UUID, authorId: UUID): ObjID
-    unsubscribeFrom(userId: UUID, authorId: UUID): Boolean
-  }
-
-  input CreateUserInput {
-    name: String
-    balance: Float
-  }
-
-  input CreatePostInput {
-    authorId: UUID
-    content: String
-    title: String
-  }
-
-  input CreateProfileInput {
-    userId: UUID
-    memberTypeId: MemberTypeId
-    isMale: Boolean
-    yearOfBirth: Int
-  }
-
-  input ChangeUserInput {
-    name: String
-  }
-
-  input ChangePostInput {
-    title: String
-  }
-
-  input ChangeProfileInput {
-    isMale: Boolean
-  }
-
-  type ObjID {
-    id: UUID
-  }
-`);
 
 export default plugin;
